refactor: Split the calendar screen into separate files
CalendarScreen.swift held the state, the week header, the episode
card, the page content and the screen itself in one 519 line file,
which is over the limit the linter enforces. Each piece now has its
own file and the largest is 241 lines.
